requirements.txt
django==3.1.2
djangorestframework==3.12.4
django-filter==21.1
djangorestframework-api-key==2.1.0
